[feature/MOV-rel32] Refactoring with separating files as mod_rm.cpp/hpp
