+++ src/mod_rm.hpp
@@ -1,0 +1,134 @@
+#ifndef MOD_RM_HPP_
+#define MOD_RM_HPP_
+
+#include <fstream>
+#include <sstream>
+#include <algorithm>
+#include <map>
+#include "nask_defs.hpp"
+
+namespace nask_utility {
+     namespace ModRM {
+	  const std::regex rm000("(AL|AX|EAX)");
+	  const std::regex rm001("(CL|CX|ECX)");
+	  const std::regex rm010("(DL|DX|EDX)");
+	  const std::regex rm011("(BL|BX|EBX)");
+
+	  const std::regex regImm08("(AL|CL|DL|BL|AH|CH|DH|BH)");
+	  const std::regex regImm16("(AX|BX|CX|DX|SI|DI|BP|SP|IP|FLAGS|CS|SS|DS|ES|FS|GS)");
+	  const std::regex regImm32("(EAX|EBX|ECX|EDX|ESI|EDI|EBP|ESP|EIP|EFLAGS)");
+
+	  enum mods {
+	       REG_REG = 0, // mod=00: [レジスター+レジスター]
+	       REG_DISP8,   // mod=01: [レジスター+disp8]
+	       REG_DISP16,  // mod=10: [レジスター+disp16/32]
+	       REG          // mod=11: レジスター
+	  };
+
+	  enum reg_field {
+	       SLASH_0 = 0, // '/0'
+	       SLASH_1,     // '/1'
+	       SLASH_2,     // '/2'
+	       SLASH_3,     // '/3'
+	       SLASH_4,     // '/4'
+	       SLASH_5,     // '/5'
+	       SLASH_6,     // '/6'
+	       SLASH_7,     // '/7'
+	       SLASH_R      // '/r'
+	  };
+
+	  const std::map<enum mods, std::string> MOD_TO_STR {
+	       { mods::REG_REG    , "00"},
+	       { mods::REG_DISP8  , "01"},
+	       { mods::REG_DISP16 , "10"},
+	       { mods::REG	  , "11"}
+	  };
+
+	  // @see: https://courses.engr.illinois.edu/ece390/resources/opcodes.html
+	  // rrr : W=0 : W=1 : reg32
+	  //------------------------
+	  // 000 : AL  : AX  : EAX
+	  // 001 : CL  : CX  : ECX
+	  // 010 : DL  : DX  : EDX
+	  // 011 : BL  : BX  : EBX
+	  // 100 : AH  : SP  : ESP
+	  // 101 : CH  : BP  : EBP
+	  // 110 : DH  : SI  : ESI
+	  // 111 : BH  : DI  : EDI
+	  const std::map<std::string, std::tuple<std::string, std::string>> REGISTERS_RRR_MAP {
+	       // reg:                [rrr,   W]
+	       { "AL", std::make_tuple("000", "0") },
+	       { "CL", std::make_tuple("001", "0") },
+	       { "DL", std::make_tuple("010", "0") },
+	       { "BL", std::make_tuple("011", "0") },
+	       { "AH", std::make_tuple("100", "0") },
+	       { "CH", std::make_tuple("101", "0") },
+	       { "DH", std::make_tuple("110", "0") },
+	       { "BH", std::make_tuple("111", "0") },
+	       { "AX", std::make_tuple("000", "1") },
+	       { "CX", std::make_tuple("001", "1") },
+	       { "DX", std::make_tuple("010", "1") },
+	       { "BX", std::make_tuple("011", "1") },
+	       { "SP", std::make_tuple("100", "1") },
+	       { "BP", std::make_tuple("101", "1") },
+	       { "SI", std::make_tuple("110", "1") },
+	       { "DI", std::make_tuple("111", "1") },
+	       {"EAX", std::make_tuple("000", "1") },
+	       {"ECX", std::make_tuple("001", "1") },
+	       {"EDX", std::make_tuple("010", "1") },
+	       {"EBX", std::make_tuple("011", "1") },
+	       {"ESP", std::make_tuple("100", "1") },
+	       {"EBP", std::make_tuple("101", "1") },
+	       {"ESI", std::make_tuple("110", "1") },
+	       {"EDI", std::make_tuple("111", "1") }
+	  };
+
+
+	  // This specify part of r/m
+	  // mmm : Function
+	  //-----------------------
+          // 000 : DS:[BX+SI]:[EAX]
+	  // 001 : DS:[BX+DI]:[ECX]
+	  // 010 : SS:[BP+SI]:[EDX]
+	  // 011 : SS:[BP+DI]:[EBX]
+          // 100 : DS:[SI]   :
+          // 101 : DS:[DI]   :
+          // 110 : SS:[BP]   :[ESI]
+	  // 111 : DS:[BX]   :[EDI]
+	  const std::map<std::string, std::string> REGISTERS_MMM_MAP {
+	       { "DS", "000"}, { "[BX+SI]", "000"}, { "[EAX]", "000"},
+	       { "DS", "001"}, { "[BX+DI]", "001"}, { "[ECX]", "001"},
+	       { "SS", "010"}, { "[BP+SI]", "010"}, { "[EDX]", "010"},
+	       { "SS", "011"}, { "[BP+DI]", "011"}, { "[EBX]", "011"},
+	       { "DS", "100"}, { "[SI]",    "100"}, // { "[]", "100"},
+	       { "DS", "101"}, { "[DI]",    "101"}, // { "[]", "101"},
+	       { "SS", "110"}, { "[BP]",    "110"}, { "[ESI]", "110"},
+	       { "DS", "111"}, { "[BX]",    "111"}, { "[EDI]", "111"},
+	  };
+
+	  // sss : Segment Register
+	  //-----------------------
+	  // 000 : ES
+	  // 001 : CS
+	  // 010 : SS
+	  // 011 : DS
+	  // 100 : FS (Only 386+)
+	  // 101 : GS (Only 386+)
+	  const std::map<std::string, std::string> SEGMENT_REGISTERS_SSS_MAP {
+	       { "ES", "000"},
+	       { "CS", "001"},
+	       { "SS", "010"},
+	       { "DS", "011"},
+	       { "FS", "100"},
+	       { "GS", "101"}
+	  };
+
+	  const std::string get_rm_from_reg(const std::string& src_reg);
+	  const uint8_t get_opecode_from_reg(uint8_t base, const std::string& reg_name);
+	  const std::string get_MMMSSS_from_reg(const std::string& reg);
+	  uint8_t generate_modrm(enum mods m, const std::string& dst_reg, enum reg_field reg = SLASH_R);
+	  bool is_accumulator(const std::string& reg);
+     };
+};
+
+#endif /* MOD_RM_HPP_ */
